+++ src/jvmMain/kotlin/com/sunnychung/application/multiplatform/hellohttp/ux/local/AppFont.kt
@@ -1,8 +1,13 @@
 package com.sunnychung.application.multiplatform.hellohttp.ux.local
 
 import androidx.compose.runtime.compositionLocalOf
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.platform.Font
 import androidx.compose.ui.unit.TextUnit
 import androidx.compose.ui.unit.sp
+import com.sunnychung.application.multiplatform.hellohttp.AppContext
+import com.sunnychung.application.multiplatform.hellohttp.manager.AppRes
 
 data class AppFont(
     val bodyFontSize: TextUnit,
@@ -19,6 +24,8 @@ data class AppFont(
     val largeInfoSize: TextUnit,
 
     val contextMenuFontSize: TextUnit,
+
+    val monospaceFontFamily: FontFamily,
 )
 
 val LocalFont = compositionLocalOf { regularFont() }
@@ -30,10 +37,23 @@ internal fun regularFont() = AppFont(
     supplementSize = 11.sp,
     badgeFontSize = 9.sp,
     codeEditorBodyFontSize = 13.sp,
-    codeEditorLineNumberFontSize = 12.sp,
+    codeEditorLineNumberFontSize = 11.sp,
 
     createLabelSize = 20.sp,
     largeInfoSize = 29.sp,
 
     contextMenuFontSize = 13.sp,
+
+    monospaceFontFamily = FontFamily(
+        Font(
+            identity = "PitagonSansMono-Regular",
+            getData = { AppContext.ResourceManager.getResource(AppRes.Font.PitagonSansMonoRegular) },
+            weight = FontWeight.Normal,
+        ),
+        Font(
+            identity = "PitagonSansMono-Bold",
+            getData = { AppContext.ResourceManager.getResource(AppRes.Font.PitagonSansMonoBold) },
+            weight = FontWeight.Bold,
+        )
+    ),
 )