+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Certificates in P7B (PKCS#7) format can now be imported
 
 ### Changed
+- The main monospace font has been changed to Pitagon Sans Mono and unified among all platforms
 - Importing CA certificates now imports all the certificates from an input file
 - Update the label of "Copy as PowerShell Invoke-WebRequest command" to confine supporting PowerShell version 6 or above only (there is no change to the underlying logic)
 - Inherited values in Request Editor are now showing at the bottom rather than the top